makes components reusable relates #76
